fix: cast timestamp to bigint in org backfill SELECTs
Postgres types a bare $1 in a SELECT list as text; the bigint created_at
columns rejected it. Add ::bigint casts.

+++ packages/db/scripts/backfill-orgs.ts
@@ -11,7 +11,7 @@ async function main() {
   await query(
     `INSERT INTO organizations (id, name, type, plan, created_by, created_at)
      SELECT DISTINCT 'org_' || a.user_id,
-            COALESCE(u.name, u.email, 'My workspace'), 'individual', 'free', a.user_id, $1
+            COALESCE(u.name, u.email, 'My workspace'), 'individual', 'free', a.user_id, $1::bigint
      FROM accounts a LEFT JOIN "user" u ON u.id = a.user_id
      WHERE a.user_id IS NOT NULL
      ON CONFLICT (id) DO NOTHING`,
@@ -20,7 +20,7 @@ async function main() {
 
   await query(
     `INSERT INTO org_members (org_id, user_id, role, created_at)
-     SELECT DISTINCT 'org_' || user_id, user_id, 'owner', $1
+     SELECT DISTINCT 'org_' || user_id, user_id, 'owner', $1::bigint
      FROM accounts WHERE user_id IS NOT NULL
      ON CONFLICT (org_id, user_id) DO NOTHING`,
     [now]
